Download and setup wincache extension for instances. Extensions list in actual php.ini file rather than in Packager. Closes #1

+++ manage/download.php
@@ -8,18 +8,22 @@ $target = $_SERVER['argv'][2];
 // download php zip archive
 $source = file_get_contents($file);
 
-$zip = new ZipArchive();
+if (strpos($file, '.zip') !== false) {
+    $zip = new ZipArchive();
 
-$temp_file = tempnam(sys_get_temp_dir(), 'PHP');
-file_put_contents($temp_file, $source);
+    $temp_file = tempnam(sys_get_temp_dir(), 'PHP');
+    file_put_contents($temp_file, $source);
 
-// open archive 
-if ($zip->open($temp_file) !== TRUE) {
-	die ("Could not open archive");
+    // open archive 
+    if ($zip->open($temp_file) !== TRUE) {
+	    die ("Could not open archive");
+    }
+
+    $zip->extractTo($target);
+    $zip->close();
+
+    // No need for this anymore
+    unlink($temp_file);
+} else {
+    file_put_contents($target, $source);
 }
-
-$zip->extractTo($target);
-$zip->close();
-
-// No need for this anymore
-unlink($temp_file);
